feat: サブフレームでの SiteProfile 取得はしない
* クロスオリジンiframeでは取得に失敗 (origin が null)
* SiteProfile の取得は siteProfile モジュールの責務に
