+++ apps/web-ext/src/components/siteProfile/index.ts
@@ -1,3 +1,4 @@
 export * from "./events";
+export * from "./messaging";
 export * from "./SiteProfile";
 export * from "./use-site-profile";
