+++ apps/web-ext/src/components/siteProfile/use-site-profile.ts
@@ -1,9 +1,8 @@
-import { deserializeIfError } from "@originator-profile/core";
 import { SpVerifier, VerifiedSp } from "@originator-profile/verify";
 import { useParams } from "react-router";
 import useSWRImmutable from "swr/immutable";
 import { getRegistryKeys } from "../../utils/get-registry-keys";
-import { siteProfileMessenger } from "./events";
+import { fetchTabSiteProfile } from "./messaging";
 
 const key = "site-profile";
 
@@ -11,34 +10,21 @@ async function fetchVerifiedSiteProfile([, tabId]: [
   _: typeof key,
   tabId: number,
 ]): Promise<VerifiedSp> {
-  const result = await siteProfileMessenger.sendMessage(
-    "fetchSiteProfile",
-    null,
-    tabId,
-  );
-
-  // コンテントスクリプトからError型を返してもError型として取り扱ってくれない
-  // そのためエラーだった場合エラー型に復元してくれるシリアライズ処理を行っているため
-  // deserializeIfErrorにてエラー型として復元してやる
-  const parsed = deserializeIfError(result);
-
-  if (parsed instanceof Error) {
-    throw parsed;
-  }
+  const data = await fetchTabSiteProfile(tabId);
 
   const [issuer, key] = getRegistryKeys();
 
   const verifySp = SpVerifier(
     {
-      ...parsed.result,
+      ...data.result,
       originators: [
         ...import.meta.env.REGISTRY_OPS,
-        ...parsed.result.originators,
+        ...data.result.originators,
       ],
     },
     key,
     issuer,
-    parsed.origin,
+    data.origin,
   );
 
   const verifiedSp = await verifySp();
